+++ airflow/providers/elasticsearch/log/es_task_handler.py
@@ -164,9 +164,7 @@ class ElasticsearchTaskHandler(FileTaskHandler, ExternalLoggingMixin, LoggingMix
     @staticmethod
     def format_url(host: str) -> str:
         """
-        Formats the given host string to ensure it starts with 'http'.
-
-        Checks if the host string represents a valid URL.
+        Format the given host string to ensure it starts with 'http' and check if it represents a valid URL.
 
         :params host: The host string to format and check.
         """
@@ -269,7 +267,7 @@ class ElasticsearchTaskHandler(FileTaskHandler, ExternalLoggingMixin, LoggingMix
 
         offset = metadata["offset"]
         log_id = self._render_log_id(ti, try_number)
-        logs = self.es_read(log_id, offset, metadata)
+        logs = self._es_read(log_id, offset)
         logs_by_host = self._group_logs_by_host(logs)
         next_offset = offset if not logs else attrgetter(self.offset_field)(logs[-1])
         # Ensure a string here. Large offset numbers will get JSON.parsed incorrectly
@@ -330,13 +328,14 @@ class ElasticsearchTaskHandler(FileTaskHandler, ExternalLoggingMixin, LoggingMix
         # Just a safe-guard to preserve backwards-compatibility
         return log_line.message
 
-    def es_read(self, log_id: str, offset: int | str, metadata: dict) -> list | ElasticSearchResponse:
+    def _es_read(self, log_id: str, offset: int | str) -> list | ElasticSearchResponse:
         """
         Return the logs matching log_id in Elasticsearch and next offset or ''.
 
         :param log_id: the log_id of the log to read.
         :param offset: the offset start to read log from.
-        :param metadata: log metadata, used for steaming log download.
+
+        :meta private:
         """
         query: dict[Any, Any] = {
             "query": {
@@ -453,7 +452,7 @@ class ElasticsearchTaskHandler(FileTaskHandler, ExternalLoggingMixin, LoggingMix
 
     def get_external_log_url(self, task_instance: TaskInstance, try_number: int) -> str:
         """
-        Creates an address for an external log collecting service.
+        Create an address for an external log collecting service.
 
         :param task_instance: task instance object
         :param try_number: task instance try_number to read logs from.
@@ -470,7 +469,7 @@ class ElasticsearchTaskHandler(FileTaskHandler, ExternalLoggingMixin, LoggingMix
 
     def _resolve_nested(self, hit: dict[Any, Any], parent_class=None) -> type[Hit]:
         """
-        Resolves nested hits from Elasticsearch by iteratively navigating the `_nested` field.
+        Resolve nested hits from Elasticsearch by iteratively navigating the `_nested` field.
 
         The result is used to fetch the appropriate document class to handle the hit.