+++ helm_tests/airflow_core/test_worker.py
@@ -61,7 +61,7 @@ class TestWorker:
             values=values,
             show_only=["templates/workers/worker-deployment.yaml"],
         )
-        expected_result = revision_history_limit if revision_history_limit else global_revision_history_limit
+        expected_result = revision_history_limit or global_revision_history_limit
         assert jmespath.search("spec.revisionHistoryLimit", docs[0]) == expected_result
 
     def test_should_add_extra_containers(self):
@@ -723,7 +723,6 @@ class TestWorkerKedaAutoScaler:
         ],
     )
     def test_should_use_keda_query(self, query, executor, expected_query):
-
         docs = render_chart(
             values={
                 "executor": executor,