+++ tests/system/providers/google/cloud/bigquery/example_bigquery_dts.py
@@ -88,7 +88,6 @@ with DAG(
     catchup=False,
     tags=["example", "bigquery"],
 ) as dag:
-
     create_bucket = GCSCreateBucketOperator(
         task_id="create_bucket", bucket_name=BUCKET_NAME, project_id=PROJECT_ID
     )