+++ airflow/providers/apache/impala/CHANGELOG.rst
@@ -26,6 +26,18 @@
 Changelog
 ---------
 
+1.2.0
+.....
+
+.. note::
+  This release of provider is only available for Airflow 2.5+ as explained in the
+  `Apache Airflow providers support policy <https://github.com/apache/airflow/blob/main/PROVIDERS.rst#minimum-supported-version-of-airflow-for-community-managed-providers>`_.
+
+Misc
+~~~~
+
+* ``Bump min airflow version of providers (#34728)``
+
 1.1.3
 .....